refactor: update item & player systems, remove old docs
- Refactor ItemManager, ItemService, Inventory
- Update player & map services
- Remove outdated docs (KICK_OLD_SESSION, SEAORM_GUIDE, network_message_panic_fix)
- Adjust Cargo.toml & Cargo.lock

+++ src/entities/player.rs
@@ -14,7 +14,7 @@ pub struct Model {
     pub account_id: Option<i32>,
     
     pub name: String,
-    pub head: i32,
+    pub head: i16,
     pub gender: i32,
     pub have_tennis_space_ship: Option<bool>,
     pub clan_id: i32,
